requirements.txt
CTkMessagebox==2.7
customtkinter==5.2.2
google_api_python_client==2.187.0
gspread==6.2.1
pandas==2.3.3
Pillow==12.1.0
protobuf==6.33.2
screeninfo==0.8.1
tabulate==0.9.0
windows==0.1.1
